+++ src/fdl.cu
@@ -1,5 +1,4 @@
 #include "fdl.h"
-#include <cstdint>
 
 
 
@@ -7,7 +6,7 @@ const float MAX_VELOCITY_MAGNITUDE_SQR = 10.0f;
 
 
 
-#define ELEM(data, cols, i, j) (data)[(int64_t)(i) * (cols) + (j)]
+#define ELEM(data, cols, i, j) (data)[(size_t)(i) * (cols) + (j)]
